+++ src/vault_image_description/__init__.py
@@ -1,0 +1,1 @@
+"""Vault Image Description plugin"""